gaialeague_ml/requirements.txt
django==3.0.2
psycopg2==2.8.4
whitenoise==5.0.1
djangorestframework==3.11.0
markdown==3.1.1
requests==2.22.0
opencv-python==4.1.2.30